CLOCK_PROCESS_CPUTIME_ID WORKS

+++ inc/env.h
@@ -66,8 +66,8 @@ struct Env {
 	uint32_t env_ipc_value;		// Data value sent to us
 	envid_t env_ipc_from;		// envid of the sender
 	int env_ipc_perm;		// Perm of page mapping received
-	struct timespec env_time;
-	int env_time_start;
+	struct timespec env_time; // amount of time process has been running
+	long long env_time_start; // moment environment start running again
 };
 
 #endif // !JOS_INC_ENV_H
